[BE_ky]: update db, api, routes
