+++ .gitignore
@@ -1,3 +1,3 @@
 __pycache__/
 *.pyc
-db/e_sol.db
+db/e_sol.db
