+++ run.py
@@ -1,4 +1,4 @@
-from app import create_app
+from __init__ import create_app
 
 app = create_app()
